Extracts Controls into its own component Grid&Controls  - Compartmentalizes Grid and Controls into separate components.  - updateGrid function and   populated2dArray state Hook lifted up into App.js  -Lifting them up allows Grid and Controls to access  both functions as props
