+++ src/styles.css
@@ -1,13 +1,25 @@
-.App {
-  font-family: sans-serif;
-  text-align: center;
-}
 * {
   box-sizing: border-box;
   width: 100%;
   text-align: center;
 }
 
+#root{
+  /* display:inline-block; */
+  background:white;
+}
+
+.App {
+  font-family: sans-serif;
+  text-align: center;
+}
+section{
+  display:inline-block;
+width:50%;
+height:100%;
+text-align: center;
+vertical-align: middle;
+}
 input#lifeCyleRangeSlide {
   max-width: 100vw;
   width: 100%;
@@ -27,4 +39,36 @@ div .sliderBox {
   flex-direction: column;
   align-items: center;
   text-align: center;
+}
+
+
+.tooltip {
+  position: relative;
+  display: inline-block;
+  border-bottom: 1px dotted black;
+}
+
+.tooltip .tooltiptext {
+  visibility: hidden;
+  width: 400px;
+  background-color: black;
+  color: #fff;
+  text-align: center;
+  border-radius: 6px;
+  padding: 5px 0;
+  position: absolute;
+  z-index: 1;
+  bottom: 100%;
+  left: 50%;
+  right:50%;
+  margin-left: -60px;
+  font-size:.75rem;
+  /* Fade in tooltip - takes 1 second to go from 0% to 100% opac: */
+  opacity: 0;
+  transition: opacity 1s;
+}
+
+.tooltip:hover .tooltiptext {
+  visibility: visible;
+  opacity: 1;
 }
